# openApi/swagger.yaml
openapi: 3.0.0
info:
  title: 'LaserLiga API'
  description: 'All API methods on the Laser liga portal.'
  version: '1.0'
servers:
  -
    url: 'https://laserliga.cz'
paths:
...
  /api/players:
    get:
      tags:
        - Players
      summary: 'Find Users'
      description: 'This method returns users based on the provided search parameters.'
      operationId: find
      parameters:
        -
          name: search
          in: query
          description: 'Search parameter (name, code, email)'
          required: false
          schema:
            type: string
        -
          name: arena
          in: query
          description: 'Home arena filter'
          required: false
          schema:
            oneOf:
              -
                type: integer
              -
                enum:
                  - self
        -
          name: connectionType
          in: query
          description: 'Connected account type filter'
          required: false
          schema:
            type: string
        -
          name: identifier
          in: query
          description: 'Connected account identifier'
          required: false
          schema:
            type: string
        -
          name: codes
          in: query
          description: 'List of user codes'
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: 'List of players'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LigaPlayer'
    post:
      tags:
        - Players
      summary: 'Register a new player'
      description: 'Register a new player'
      operationId: registerPlayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  example: 'Lay zerteg'
                  nullable: false
                email:
                  type: string
                  format: email
                  example: lay@zerteg.com
                  nullable: false
                password:
                  type: string
                  example: superstrongpassword123
                  nullable: false
              type: object
      responses:
        '201':
          description: 'Created a new player'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LigaPlayer'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 'Internal error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/api/players/{code}':
    get:
      tags:
        - Players
      summary: 'Get User by Code'
      description: 'This method returns a user based on the provided code.'
      operationId: player
      parameters:
        -
          name: code
          in: path
          description: 'User code'
          required: true
          schema:
            type: string
            pattern: '^\d+-[A-Z\d]{5}$'
      responses:
        '200':
          description: 'Player fetched successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LigaPlayer'
        '404':
          description: 'Player not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/api/players/{code}/title':
    get:
      tags:
        - Players
      summary: "Get User's Title by Code"
      description: 'This method returns the title of a user based on the provided code.'
      operationId: playerTitle
      parameters:
        -
          name: code
          in: path
          description: 'User code'
          required: true
          schema:
            type: string
            pattern: '^\d+-[A-Z\d]{5}$'
      responses:
        '200':
          description: "Player's title fetched successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Title'
        '404':
          description: 'Player not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/api/players/old/{code}':
    get:
      tags:
        - Players
      summary: 'Find Users by old code'
      description: 'Finds players whose code has changed.'
      operationId: findByOldCode
      parameters:
        -
          name: code
          in: path
          description: 'User code'
          required: true
          schema:
            type: string
            pattern: '^\d+-[A-Z\d]{5}$'
      responses:
        '200':
          description: 'List of players'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LigaPlayer'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Title:
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        rarity:
          $ref: '#/components/schemas/Rarity'
        unlocked:
          type: boolean
      type: object
    Address:
      properties:
        street:
          type: string
          example: 'OD Luna 3. patro, Velké náměstí 175'
          nullable: true
        city:
          type: string
          example: Písek
          nullable: true
        postCode:
          type: string
          example: '39701'
          nullable: true
        country:
          type: string
          example: Česko
          nullable: true
      type: object
    Arena:
      properties:
        name:
          type: string
          example: 'Laser arena Písek'
        gameCodePrefix:
          type: string
          example: g
        lat:
          type: number
          format: float
          example: 49.307678
          nullable: true
        lng:
          type: number
          format: float
          example: 14.147773
          nullable: true
        address:
          $ref: '#/components/schemas/Address'
        web:
          type: string
          example: 'https://laserarenapisek.cz'
          nullable: true
        contactEmail:
          type: string
          example: info@laserarenapisek.cz
          nullable: true
        contactPhone:
          type: string
          example: '+420 776 606 631'
          nullable: true
      type: object
    LigaPlayer:
      description: 'Same as the regular player, but with the addition of the arena and user parameters'
      type: object
      allOf:
        -
          $ref: '#/components/schemas/LigaPlayerBase'
        -
          properties:
            user: {  }
            arena:
              oneOf:
                -
                  $ref: '#/components/schemas/Arena'
              nullable: true
          type: object
    LigaPlayerBase:
      properties:
        stats:
          $ref: '#/components/schemas/PlayerStats'
        code:
          description: 'Unique code for each player - two players can have the same code if they are from different arenas.'
          type: string
        nickname:
          type: string
        email:
          type: string
        avatar:
          type: string
          nullable: true
        avatarStyle:
          type: string
          nullable: true
        avatarSeed:
          type: string
          nullable: true
        title:
          oneOf:
            -
              $ref: '#/components/schemas/Title'
          nullable: true
        birthday:
          type: string
          format: date-time
          nullable: true
      type: object
...
    PlayerStats:
      properties:
        gamesPlayed:
          type: integer
        arenasPlayed:
          type: integer
        rank:
          type: integer
        averageAccuracy:
          type: number
          format: float
        averagePosition:
          type: number
          format: float
        maxAccuracy:
          type: integer
        maxScore:
          type: integer
        maxSkill:
          type: integer
        shots:
          type: integer
        averageShots:
          type: number
          format: float
        averageShotsPerMinute:
          type: number
          format: float
        totalMinutes:
          type: integer
        kd:
          type: number
          format: float
        hits:
          type: integer
        deaths:
          type: integer
      type: object
...
    Rarity:
      type: string
      enum:
        - common
        - uncommon
        - rare
        - epic
        - legendary
        - mythic
...
    ErrorResponse:
      properties:
        title:
          type: string
          example: 'Error title'
        type:
          $ref: '#/components/schemas/ErrorType'
        detail:
          type: string
          example: 'Error description'
          nullable: true
        exception:
          properties:
            message:
              type: string
              example: 'Some exception description'
            code:
              type: integer
              example: 123
            trace:
              type: array
              items:
                type: object
              example:
                -
                  file: index.php
                  line: 1
                  function: abc
                  args: ['Argument value']
          type: object
          nullable: true
        values:
          type: object
          example:
            key1: value1
            key2: value2
          nullable: true
      type: object
...
    ErrorType:
      type: string
      enum:
        - validation_error
        - database_error
        - internal_error
        - resource_not_found_error
        - resource_access_error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  -
    bearerAuth: []
...